feat: implement compiler options for transpile
